Moved normalize_event_filename() to utils.py
