+++ docs/technical_changes.md
@@ -2,6 +2,7 @@
 - Changed `update_token` method in `WebSettingsApi` class. No more `static.`
 - Removed `is_osxbundle()` method from `LocalClient` class
 - Removed `is_updated()` method from `Manager` class. Use `updated` attribute instead.
+- Moved `normalize_event_filename()` function from engine/watcher/local_watcher.py to utils.py
 
 # 2.5.4
 - Moved `get_repository_names()` method from `RemoteDocumentClient` class to `RemoteDocumentClientForTests` class, only for tests
